Complete MealMate workflow.
Build 3 agents(i.e. customer manager, chef de cuisine, and line cook)
and make them communicate with each based on defined prompt.

Dataflow between them is basically a list of dict with ingredients name
and description.

Next step is to work on availability check function based on the list,
and connect with the database.